# src/docs/blocktimestamp/swagger.yaml
definitions:
  integrated.Result:
    properties:
      account:
        type: string
      balance:
        type: string
      balanceHex:
        type: string
      tokenAddress:
        type: string
      transfer_history:
        items:
          $ref: '#/definitions/integrated.TransferHistory'
        type: array
    type: object
  integrated.TransferHistory:
    properties:
      from:
        type: string
      timestamp:
        type: integer
      to:
        type: string
      txHash:
        type: string
      value:
        type: string
      valueHex:
        type: string
    type: object
...
  router.ErrorResponse:
    properties:
      error:
        example: Description of the error
        type: string
    type: object
...
  router.ResponseBlockTimestamp:
    properties:
      date:
        description: Original date
        example: '2024-01-01'
        type: string
      hexTimestamp:
        description: Hexadecimal timestamp
        example: '0x62d43b1f'
        type: string
      timestamp:
        description: Unix timestamp
        example: 1672531199
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /get/getBlockTimestamp:
    get:
      description:
        Fetches the block timestamp for a specific block number in both
        Unix and hexadecimal formats.
      parameters:
        - description: Block number in decimal or hexadecimal format
          in: query
          name: number
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/router.ResponseBlockTimestamp'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Retrieve block timestamp by block number
      tags:
        - Block
...
  /track/trackERC20:
    get:
      description:
        Tracks ERC20 token transfers for a specific account and token contract
        within a given date and target count.
      parameters:
        - description: Ethereum account address (e.g., '0x1234567890abcdef1234567890abcdef12345678')
          in: query
          name: account
          required: true
          type: string
        - description: ERC20 token contract address (e.g., '0xabc123')
          in: query
          name: tokenAddress
          required: true
          type: string
        - description: Date in 'YYYY-MM-DD' format
          in: query
          name: date
          required: true
          type: string
        - description: Number of transactions to retrieve
          in: query
          name: targetCount
          type: integer
        - description: Time limit for processing in seconds
          in: query
          name: timeLimit
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/integrated.Result'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Track ERC20 token transfers
      tags:
        - Track
...
swagger: '2.0'